requirements.txt
backoff==2.2.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
Flask==2.3.1
Flask-OpenTracing==1.1.0
googleapis-common-protos==1.56.2
grpcio==1.54.2
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
jaeger-client==4.8.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
opentelemetry-api==1.18.0
opentelemetry-exporter-jaeger-proto-grpc==1.18.0
opentelemetry-exporter-jaeger-thrift==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
opentracing==2.4.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
protobuf==3.20.3
requests==2.31.0
six==1.16.0
threadloop==1.0.2
thrift==0.16.0
tornado==6.3.1
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.2
wrapt==1.15.0
zipp==3.15.0
